fix: headers being incorrectly matched in the gRPC metadata
