+++ code/go/0chain.net/blobbercore/handler/grpc_commit_handler.go
@@ -28,7 +28,7 @@ func (b *blobberGRPCService) Commit(ctx context.Context, req *blobbergrpc.Commit
 	if err != nil {
 		return nil, err
 	}
-	httpRequestWithMetaData(r, GetGRPCMetaDataFromCtx(ctx), req.Allocation)
+	httpRequestWithMetaData(r, getGRPCMetaDataFromCtx(ctx), req.Allocation)
 	r.Header.Set("Content-Type", writer.FormDataContentType())
 
 	resp, err := CommitHandler(ctx, r)
