--- code/go/0chain.net/blobbercore/handler/helper.go
@@ -1,20 +1,0 @@
-package handler
-
-import (
-	"context"
-
-	blobbergrpc "github.com/0chain/blobber/code/go/0chain.net/blobbercore/blobbergrpc/proto"
-	"github.com/gorilla/mux"
-	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
-	"google.golang.org/grpc"
-)
-
-func registerGRPCServices(r *mux.Router, server *grpc.Server) {
-	blobberService := newGRPCBlobberService()
-	grpcGatewayHandler := runtime.NewServeMux()
-
-	blobbergrpc.RegisterBlobberServiceServer(server, blobberService)
-	_ = blobbergrpc.RegisterBlobberServiceHandlerServer(context.Background(), grpcGatewayHandler, blobberService)
-	r.PathPrefix("/").Handler(grpcGatewayHandler)
-
-}
